fix:move code for place images slider to its new component
