+++ src/components/place-page/PlaceDetails.js
@@ -6,7 +6,7 @@ const PlaceDetails = ({ place }) => {
         <article className="c-white overpass">
             <h1 className="medium-text bold">Description</h1>
             <p className="small-text">{place.description}</p>
-            {place.number === '' ? null
+            {place.phone === '' ? null
               : <p className="small-text">
                     <span><BiPhone className="c-green medium-text" /> </span>{place.phone}
                 </p> }
