+++ src/components/place-page/PlacePage.js
@@ -1,15 +1,14 @@
 import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
-import { FaFacebook, FaInstagramSquare } from 'react-icons/fa';
+import { FaFacebookSquare, FaInstagramSquare } from 'react-icons/fa';
 import { checkLoading } from '../../state/sightseeing/sightseeing.actions';
 import NavBar from '../Navbar/NavBar';
 import PageIntro from '../PageIntro';
 import PreLoader from '../pre-loader/PreLoader';
-import PlaceImage from './PlaceImage';
-import CustomSlider from '../slider/CustomSlider';
 import PlaceDetails from './PlaceDetails';
 import PlaceMap from './PlaceMap';
 import SocilaMediaIcon from './SocilaMediaIcon';
+import PlaceImagesSlider from './PlaceImagesSlider';
 
 const PlacePage = ({ checkLoading, loading }) => {
   const [place, setPlace] = useState({});
@@ -19,19 +18,6 @@ const PlacePage = ({ checkLoading, loading }) => {
     setImages(JSON.parse(localStorage.getItem('placeImages')));
     checkLoading(false);
   };
-  const sliderFor = {
-    dots: true,
-    infinite: true,
-    speed: 500,
-    slidesToShow: 1,
-    slidesToScroll: 1,
-    autoplaySpeed: 3000,
-    autoplay: true,
-    pauseOnHover: false,
-    lazyLoad: 'ondemand',
-    centerMode: true
-  };
-
   useEffect(() => {
     loadPage();
   }, []);
@@ -44,17 +30,7 @@ const PlacePage = ({ checkLoading, loading }) => {
                 <NavBar bg="fade-black c-white" />
                 <PageIntro title={place.name} color="c-white mt-2" />
             </header>
-            <div className="container-fluid">
-                {images === [] ? null : (
-                    <>
-                    <CustomSlider customStyles={sliderFor} nextAndPrev="true">
-                        {images.map((key) => (
-                            <PlaceImage image={key} key={key} />
-                        ))}
-                    </CustomSlider>
-                    </>
-                )}
-            </div>
+            <PlaceImagesSlider images={images} />
             <div className="two-sec-grid-cols">
                 <div className="left-col p-2">
                     <PlaceMap place={place} />
@@ -62,7 +38,7 @@ const PlacePage = ({ checkLoading, loading }) => {
                 <div className="right-col p-2 center column text-left">
                     <PlaceDetails place={place} />
                     <div className="left w-100 text-left">
-                      {place.facebook === '' ? null : <SocilaMediaIcon icon={<FaFacebook />} link={place.facebook} />}
+                      {place.facebook === '' ? null : <SocilaMediaIcon icon={<FaFacebookSquare />} link={place.facebook} />}
                       {place.instagram === '' ? null : <SocilaMediaIcon icon={<FaInstagramSquare />} link={place.instagram} />}
                     </div>
                 </div>
