--- src/pages/about.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import SEO from '../components/seo';
-import Layout from '../components/layout';
-import NavBar from '../components/Navbar/NavBar';
-
-const AboutPage = () => (
-  <Layout>
-    <SEO title="About" />
-    <NavBar />
-  </Layout>
-);
-
-export default AboutPage;
